Change type _period field from uint16_t to uint32_t.

+++ lib/Timer_Blink/timer_blink.cpp
@@ -1,7 +1,7 @@
 #include "timer_blink.h"
 
 /*-----------------------------| Class Timer |-------------------------------*/
-Timer::Timer(uint16_t period)  : _period(period) {}
+Timer::Timer(uint32_t period)  : _period(period) {}
 
 Timer::~Timer() {}
 
@@ -13,13 +13,13 @@ bool Timer::ready() {
       return false;
 }
 
-void Timer::setPeriod(uint16_t period) { _period = period; }
+void Timer::setPeriod(uint32_t period) { _period = period; }
 void Timer::resetCounter() { _tmr = millis(); }
-uint16_t Timer::getPeriod() { return _period; }
+uint32_t Timer::getPeriod() { return _period; }
 uint32_t Timer::getTmr() { return _tmr; }
 
 /*-----------------------------| Class Blink |-------------------------------*/
-Blink::Blink(uint16_t period) : Timer(period) {}
+Blink::Blink(uint32_t period) : Timer(period) {}
   
 Blink::~Blink() {}
